requirements.txt
requests
pycryptodome
py-bcrypt
six
